plugin: add pre, post, and save directives
Before this commit: pre and post commands are executed as regular
cli commands. The state of the execution is captured. If `on_error`
option is set to `continue`, the plugin continues the it execution
when encountering errors.

After this commit: configuration file has `pre`, `post`, and `saveas`
directives. If `on_error` option is set to `continue` and error is
triggered during `pre` stage, the plugin will not continue.

Smoke-test: build a container and run the plugin against a firewall.