various: major upgrade of the entire analytics engine
Before this commit: the analytics engine and expect is not in sync and
there is no proper exception handling and logging.

After this commit: the analytics engine and expect communication over
pipes is handled better and includes proper exception logging.

Smoke-test: build a container and run the plugin against a firewall,
virtual switches, and linux servers.

More info: the commit includes improvements to the default cli command
sets. A conditions framework is added as well, e.g. a command will not
run unless it satisfies a condition precedent, etc.